Fixing anonymous commenter issue and refactoring navbar to navigate back to the correct selectors
